# Simplify and 3x Speedup CausalDotProduct CUDA Kernel for Larger Hidden Sizes