### Summary

This branch optimizes the `causal_dot_product_` and `causal_dot_backward_` functions and related kernels for models with larger hidden sizes.

### Motivations

#77 made the CUDA kernels for causal product much faster, especially for the benchmarks where the head `H=8` and hidden sizes `E = M = 64` as in the paper. However, it **falls back** to the original implementation when the hidden size is larger, e.g., using one head (N=16, L=4096, H=1, E=M=512), and becomes significant slower and hard to scale up.

To tackle this issue, this commit optimizes the original (fallback) implementation. This new code uses much less `sync_threads`s and `atomicAdd`s and less shared memory.

### Benchmark Results

To profile its results in more extensive settings, use the following code:

```python
def _test_benchmark_forward(self, CP):
    print("{:>4} {:>5} {:>5} {:>5} {:>5} {:>12}".format("N", "L", "H", "E", "M", "FW Time (ms)"))
    for N, L, H, E, M in [
        [8, 4096, 8, 512, 512],
        [1, 4096, 8, 512, 512],
        [16, 4096, 1, 512, 512],
        [16, 4096, 8, 256, 256],
        [16, 4096, 8, 128, 128],
        [16, 4096, 8, 64, 64],
    ]:
        Q = torch.rand(N, H, L, E).cuda()
        K = torch.rand(N, H, L, E).cuda()
        V = torch.rand(N, H, L, M).cuda()
        out = torch.rand(N, H, L, M).cuda()

        # warmup the cache
        for i in range(10):
            self.kernels[CP]["forward"](Q, K, V, out)

        # measure
        start = torch.cuda.Event(enable_timing=True)
        end = torch.cuda.Event(enable_timing=True)
        start.record()
        for i in range(10):
            self.kernels[CP]["forward"](Q, K, V, out)
        end.record()
        torch.cuda.synchronize()
        print("{:>5} {:>5} {:>5} {:>5} {:>5} {:>8.1f}".format(
            N, L, H, E, M,
            start.elapsed_time(end)
        ))
```

#### Forward Pass

* New

```
    N     L     H     E     M Time (ms)
    8  4096     8   512   512   2541.4
    1  4096     8   512   512    311.2
   16  4096     1   512   512    630.6
   16  4096     8   256   256   1235.1
   16  4096     8   128   128     65.2
   16  4096     8    64    64     20.0
```

* Old

```
    N     L     H     E     M Time (ms)
    8  4096     8   512   512   7863.3
    1  4096     8   512   512   1008.9
   16  4096     1   512   512   1988.6
   16  4096     8   256   256   3666.2
   16  4096     8   128   128     63.1
   16  4096     8    64    64     20.9
```

#### Backward Pass

* New

```
   N     L     H     E     M Time (ms)
    8  4096     8   512   512   8765.8
    1  4096     8   512   512    978.9
   16  4096     1   512   512   1981.6
   16  4096     8   256   256   4507.5
   16  4096     8   128   128    180.9
   16  4096     8    64    64     52.3
```

* Old

```
   N     L     H     E     M Time (ms)
    8  4096     8   512   512  29759.8
    1  4096     8   512   512   3938.9
   16  4096     1   512   512   7775.5
   16  4096     8   256   256  11615.8
   16  4096     8   128   128    184.1
   16  4096     8    64    64     54.1
```